+++ packages/wise_nav_bar/lib/src/utils/utils.dart
@@ -1,2 +1,2 @@
-export 'platform_nav_bar_helper.dart';
 export 'platform_scaffold.dart';
+export 'platform_widget_helper.dart';