+++ packages/wise_nav_bar/CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.0.1
 
-* Different implementations of native App/Navigation bars
+* Base PlatformAppBar (with guidelines)
+* Text AppBar